reenable warnings

+++ src/include/arch/memmgt.h
@@ -37,6 +37,6 @@ mmap_page_t *find_best_page(uint32_t size);
 
 mmap_page_t *get_block_page(mmap_block_t *block);
 
-mmap_block_t *get_block_of_ptr(uint64_t addr);
+mmap_block_t *get_block_of_ptr(uint32_t addr);
 
 #endif //GOOSE_MEMMGT_H
